GameCenter Fix 2nd attempt
Kept the buttons I created in 1st attempt and used https://forum.unity.com/threads/tutorial-gamecenter-leaderboards-in-5-minutes-for-intermediate-advanced-scripters.238824/

+++ Assets/game1.cs
@@ -1,7 +1,6 @@
 using UnityEngine;
 using UnityEngine.SocialPlatforms;  
 using UnityEngine.UI; //add when refrencing textbox in unity   
-
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading;
@@ -11,13 +10,13 @@ using UnityEngine.SceneManagement;
 public class game1 : MonoBehaviour
 {
 
- int scorea;
+ long scorea; // Gamecenter requires a long variable
  
    public void Who() 
     { 
           scorea = Score.scoreValue;  
                 Social.ReportScore(scorea,"com.dvf.leaderboard",HighScoreCheck);
-
+                
     }
 
      static void HighScoreCheck(bool result)
